+++ projects/packages/my-jetpack/src/class-initializer.php
@@ -43,7 +43,7 @@ class Initializer {
 	 *
 	 * @var string
 	 */
-	const PACKAGE_VERSION = '5.42.3';
+	const PACKAGE_VERSION = '5.43.1';
 
 	/**
 	 * HTML container ID for the IDC screen on My Jetpack page.